# Split lib/benchmark/Prelude to decrease memory consumed during compilation