--- a/bench.sh
+++ b/bench.sh
@@ -9,8 +9,8 @@ base_stream_grp="\
     Data.Stream.StreamDK"
 SERIAL_O_1="linear"
 SERIAL_O_n="serial-o-n-heap serial-o-n-stack serial-o-n-space"
-UNFOLD_BENCHMARKS="unfold-o-1-space unfold-o-n-space"
 serial_grp="$SERIAL_O_1 $SERIAL_O_n"
+
 # parallel benchmark-suite is separated because we run it with a higher
 # heap size limit.
 concurrent_grp="linear-async linear-rate nested-concurrent parallel concurrent adaptive"
@@ -40,7 +40,8 @@ all_grp="\
     $serial_grp \
     $concurrent_grp \
     $array_grp \
-    $parser_grp"
+    $parser_grp \
+    Data.Unfold"
 
 ALL_BENCH_GROUPS="\
     all_grp \
@@ -55,8 +56,6 @@ ALL_BENCH_GROUPS="\
 # RTS options that go inside +RTS and -RTS while running the benchmark.
 bench_rts_opts () {
   case "$1" in
-    "unfold-o-1-space") echo -n "-T -K36K -M16M" ;;
-    "unfold-o-n-space") echo -n "-T -K32M -M64M" ;;
     "linear") echo -n "-T -K36K -M16M" ;;
     "serial-o-n-stack") echo -n "-T -K1M -M16M" ;;
     "serial-o-n-heap") echo -n "-T -K36K -M128M" ;;
@@ -74,10 +73,6 @@ bench_rts_opts () {
 # The correct executable for the given benchmark name.
 bench_exec () {
   case "$1" in
-    "fold-o-1-space") echo -n "fold" ;;
-    "fold-o-n-heap") echo -n "fold" ;;
-    "unfold-o-1-space") echo -n "unfold" ;;
-    "unfold-o-n-space") echo -n "unfold" ;;
     "linear") echo -n "serial" ;;
     "serial-o-n-stack") echo -n "serial" ;;
     "serial-o-n-heap") echo -n "serial" ;;
@@ -89,10 +84,6 @@ bench_exec () {
 # Specific gauge options for the given benchmark.
 bench_gauge_opts () {
   case "$1" in
-    "fold-o-1-space") echo -n "-m prefix o-1-space" ;;
-    "fold-o-n-heap") echo -n "-m prefix o-n-heap" ;;
-    "unfold-o-1-space") echo -n "-m prefix o-1-space" ;;
-    "unfold-o-n-space") echo -n "-m prefix o-n-space" ;;
     "linear") echo -n "-m prefix o-1-space" ;;
     "serial-o-n-stack") echo -n "-m prefix o-n-stack" ;;
     "serial-o-n-heap") echo -n "-m prefix o-n-heap" ;;