--- a/streamly.cabal
+++ b/streamly.cabal
@@ -144,7 +144,6 @@ extra-source-files:
     benchmark/Streamly/Benchmark/Data/Stream/*.hs
     benchmark/Streamly/Benchmark/FileIO/*.hs
     benchmark/Streamly/Benchmark/Prelude/*.hs
-    benchmark/Streamly/Benchmark/Prelude/Serial/*.hs
 
 extra-tmp-files:
     config.log