--- a/streamly.cabal
+++ b/streamly.cabal
@@ -102,6 +102,7 @@ extra-source-files:
     .hlint.yaml
     .travis.yml
     appveyor.yml
+    bin/bench-exec-one.sh
     cabal.project.ci
     Changelog.md
     charts-0/streamly-vs-list-time.svg
@@ -140,6 +141,7 @@ extra-source-files:
     benchmark/lib/Streamly/Benchmark/*.hs
     benchmark/Streamly/Benchmark/Memory/*.hs
     benchmark/Streamly/Benchmark/Data/*.hs
+    benchmark/Streamly/Benchmark/Data/Parser/*.hs
     benchmark/Streamly/Benchmark/Data/Prim/*.hs
     benchmark/Streamly/Benchmark/Data/Stream/*.hs
     benchmark/Streamly/Benchmark/FileIO/*.hs