… .hlint.yaml
… benchmark/Chart.hs
… benchmark/README.md
… benchmark/Streamly/Benchmark/Data/Array.hs
… benchmark/Streamly/Benchmark/Data/Fold.hs
… benchmark/Streamly/Benchmark/Data/Parser.hs
… benchmark/Streamly/Benchmark/Data/Parser/ParserD.hs
… benchmark/Streamly/Benchmark/Data/Parser/ParserK.hs
… benchmark/Streamly/Benchmark/Data/Prim/Array.hs
… benchmark/Streamly/Benchmark/Data/SmallArray.hs
… benchmark/Streamly/Benchmark/Data/Stream/BaseStreams.hs
… benchmark/Streamly/Benchmark/Data/Stream/StreamD.hs
… benchmark/Streamly/Benchmark/Data/Stream/StreamDK.hs
… benchmark/Streamly/Benchmark/Data/Stream/StreamK.hs
… benchmark/Streamly/Benchmark/Data/Unfold.hs
… benchmark/Streamly/Benchmark/Memory/Array.hs
… benchmark/Streamly/Benchmark/Prelude/Ahead.hs
… benchmark/Streamly/Benchmark/Prelude/Async.hs
… benchmark/Streamly/Benchmark/Prelude/Concurrent.hs
… benchmark/Streamly/Benchmark/Prelude/LinearAsync.hs
… benchmark/Streamly/Benchmark/Prelude/LinearRate.hs
… benchmark/Streamly/Benchmark/Prelude/NestedConcurrent.hs
… benchmark/Streamly/Benchmark/Prelude/NestedOps.hs
… benchmark/Streamly/Benchmark/Prelude/Parallel.hs
… benchmark/Streamly/Benchmark/Prelude/Rate.hs
… benchmark/Streamly/Benchmark/Prelude/Serial.hs
… benchmark/Streamly/Benchmark/Prelude/Serial/O_1_Space.hs
… benchmark/Streamly/Benchmark/Prelude/Serial/O_n_Heap.hs
… benchmark/Streamly/Benchmark/Prelude/Serial/O_n_Space.hs
… benchmark/Streamly/Benchmark/Prelude/Serial/O_n_Stack.hs