Closes #490