@adithyaov 2020-05-08
The peak mempry usage for `cabal v2-build --enable-tests --enable-benchmarks --flag debug all -w /opt/ghc/8.8.1/bin/ghc` is about 2.5GB