@harendra-kumar 2020-06-07
Some perf difference observations after the change:

* The lowest timings earlier were 32 us, now mostly 42 us in serial stream benchmarks
* serial zip benchmarks seem to have regressed from 110 ms to 126 ms.
* asyncly/toNull 29 ms to 73 ms, and similarly most other monad benchmarks
* tapAsync/tapAsyncS seem to have regressed

rest all look the same.