@harendra-kumar 2020-06-08
the `tapAsync` benchmark change is because we now run it with threaded RTS and using two CPUs. Other benchmark regressions I have fixed, so we are now mostly at par with earlier benchmark results.